reader-bot/bot/requirements.txt
python-telegram-bot
psycopg2-binary
alembic